+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 ## Added
+- Add `DiskCacheBuilder::set_sync_to_disk_on_cache_change` to specify that the cache changes should be written to disk on every cache change.
+- Add `sync_to_disk_on_cache_change` to `#[io_cached]` to allow setting `DiskCacheBuilder::set_sync_to_disk_on_cache_change` from the proc macro.
+- Add `DiskCacheBuilder::set_connection_config` to give more control over the sled connection.
+- Add `connection_config` to `#[io_cached]` to allow setting `DiskCacheBuilder::set_connection_config` from the proc macro.
+- Add `DiskCache::connection()` and `DiskCache::connection_mut()` to give access to the underlying sled connection.
 ## Changed
 - [Breaking] `type` attribute is now `ty`
 - Upgrade to syn2